Updated for OSBA

+++ scripts/init-azure-mysql-table.sh
@@ -1,10 +1,17 @@
 #!/bin/sh
 set -e -x
 
-MYSQL_USER="<mysql-user>"
-MYSQL_PASSWORD="<mysql-password>"
-MYSQL_HOST="<mysql-host>"
-MYSQL_DB="<mysql-db>"
-MYSQL_QUERY="DROP TABLE IF EXISTS azurevote; CREATE TABLE azurevote (voteid INT NOT NULL AUTO_INCREMENT,votevalue VARCHAR(45) NULL,PRIMARY KEY (voteid));"
+if [ $# -ne 1 ]
+then
+    echo "USAGE: $0 <osba-secret-name>"
+    exit 1
+fi
+OSBA_SECRET=$1
+
+user=$(kubectl get secret $OSBA_SECRET -o jsonpath="{.data.username}" | base64 --decode)
+password=$(kubectl get secret $OSBA_SECRET -o jsonpath="{.data.password}" | base64 --decode)
+host=$(kubectl get secret $OSBA_SECRET -o jsonpath="{.data.host}" | base64 --decode)
+db="$(kubectl get secret $OSBA_SECRET -o jsonpath="{.data.database}" | base64 --decode)"
+query="DROP TABLE IF EXISTS azurevote; CREATE TABLE azurevote (voteid INT NOT NULL AUTO_INCREMENT,votevalue VARCHAR(45) NULL,PRIMARY KEY (voteid));"
 kubectl run mysql-client --image=mysql:5.7 -it --rm --restart=Never --\
-    bash -ic "mysql -u $MYSQL_USER -h $MYSQL_HOST -p$MYSQL_PASSWORD $MYSQL_DB -e\"$MYSQL_QUERY\""
+    bash -ic "mysql -u $user -h $host -p$password $db -e\"$query\""
